Fix Jacobian exercises.
* exercises/05-jacobian/jacobian.arts:

Adapt the controfile to latest ARTS version (arts-2-3-983 and above).

The new implementation of `iyEmissionStandard` introduced in
arts-2-3-983 requires a modification of the controlfile. Absorption
coefficients as a function of frequency and pressure are no longer
available as auxiliary output. A customized agenda writes these
variables to an XML file to circumvent this issue.

* exercises/05-jacobian/plot_jacobian.py:

Adapt the plotting script to work with new ARTS output.

Minor modifications (e.g. visual guides)of the created plots.

* exercises/05-jacobian/plot_jacobian.m:

Remove inoperative Matlab plotting script.

* 05-jacobian/exe5_description.{docx,pdf}:

Remove mention of the Matlab plotting script.

git-svn-id: https://arts.mi.uni-hamburg.de/svn/rt/arts-lectures/trunk@11105 aaf1aab0-4228-0410-ad68-8dceda47f409